requirements.txt
fastapi==0.95.2
cx_Oracle==8.3.0
python-dotenv==1.0.0
uvicorn==0.22.0
Jinja2==3.1.2